+++ flake8_trio/visitors/__init__.py
@@ -17,9 +17,9 @@ default_disabled_error_codes: list[str] = []
 
 # Import all visitors so their decorators run, filling the above containers
 # This has to be done at the end to avoid circular imports
-from . import visitor2xx  # isort: skip
 from . import visitor102  # isort: skip
 from . import visitor103_104  # isort: skip
-from . import visitor107_108  # isort: skip
 from . import visitor111  # isort: skip
+from . import visitor2xx  # isort: skip
+from . import visitor91x  # isort: skip
 from . import visitors  # isort: skip