+++ tests/eval_files/trio910.py
@@ -10,7 +10,7 @@ def __() -> Any:
     ...
 
 
-# ARG --enable-visitor-codes-regex=(TRIO107)|(TRIO108)
+# ARG --enable-visitor-codes-regex=(TRIO910)|(TRIO911)
 
 
 # function whose body solely consists of pass, ellipsis, or string constants is safe